requirements.txt
python-dotenv

streamlit
streamlit-chat
streamlit-authenticator

gtts
ollama
openai
mistralai
